inel: added split_buffered method
