+++ crates/inel/src/io/mod.rs
@@ -11,4 +11,4 @@ pub use buffered::{BufReader, BufWriter, FixedBufReader, FixedBufWriter};
 pub use owned::{AsyncReadOwned, AsyncWriteOwned};
 pub use split::{ReadHandle, WriteHandle};
 
-pub(crate) use split::split;
+pub(crate) use split::{split, split_buffered};
