+++ crates/inel/tests/unit/fs/file.rs
@@ -1,7 +1,8 @@
 use std::os::fd::{FromRawFd, IntoRawFd};
 
 use crate::helpers::*;
-use inel::io::AsyncWriteOwned;
+use futures::{AsyncReadExt, AsyncWriteExt};
+use inel::io::{AsyncReadOwned, AsyncWriteOwned};
 
 #[test]
 fn create() {
@@ -146,6 +147,96 @@ fn metadata() {
 
     std::fs::remove_file(&file).unwrap();
     std::fs::remove_dir(&dir).unwrap();
+}
+
+#[test]
+fn split() {
+    setup_tracing();
+
+    let name = temp_file();
+    let name_clone = name.clone();
+
+    inel::block_on(async move {
+        let file = inel::fs::File::options()
+            .readable(true)
+            .writable(true)
+            .create(true)
+            .open(name_clone)
+            .await
+            .unwrap();
+
+        let (mut reader, mut writer) = file.split();
+
+        let read = inel::spawn(async move {
+            let (buf, res) = reader.read_owned(Box::new([0; 256])).await;
+            assert!(res.is_ok_and(|r| r == 0));
+            assert_eq!(buf, Box::new([0; 256]));
+
+            let res = reader.metadata().await;
+            assert!(res.is_ok());
+
+            true
+        });
+
+        let write = inel::spawn(async move {
+            let (buf, res) = writer.write_owned(Box::new([b'A'; 512])).await;
+            assert!(res.is_ok_and(|w| w == 512));
+            assert_eq!(buf, Box::new([b'A'; 512]));
+
+            let res = writer.sync_data().await;
+            assert!(res.is_ok());
+
+            true
+        });
+
+        assert_eq!(read.join().await, Some(true));
+        assert_eq!(write.join().await, Some(true));
+    });
+
+    std::fs::remove_file(&name).unwrap();
+}
+
+#[test]
+fn split_buffered() {
+    setup_tracing();
+
+    let name = temp_file();
+    let name_clone = name.clone();
+
+    inel::block_on(async move {
+        let file = inel::fs::File::options()
+            .readable(true)
+            .writable(true)
+            .create(true)
+            .open(name_clone)
+            .await
+            .unwrap();
+
+        let (mut reader, mut writer) = file.split_buffered();
+
+        let read = inel::spawn(async move {
+            let mut buf = Box::new([0; 256]);
+            let res = reader.read(&mut buf[..]).await;
+            assert!(res.is_ok_and(|r| r == 0));
+            assert_eq!(buf, Box::new([0; 256]));
+
+            true
+        });
+
+        let write = inel::spawn(async move {
+            let buf = Box::new([b'A'; 512]);
+            let res = writer.write(&buf[..]).await;
+            assert!(res.is_ok_and(|w| w == 512));
+            assert_eq!(buf, Box::new([b'A'; 512]));
+
+            true
+        });
+
+        assert_eq!(read.join().await, Some(true));
+        assert_eq!(write.join().await, Some(true));
+    });
+
+    std::fs::remove_file(&name).unwrap();
 }
 
 #[test]
